* Scaffold/scaffold.cc (main): Fix a compiler warning: call to 'pow' is ambiguous
git-svn-id: https://svn.bcgsc.ca/svn/abyss/trunk@76249 ad02020f-2a26-4133-947e-1829317c1416

+++ Scaffold/scaffold.cc
@@ -728,7 +728,7 @@ int main(int argc, char** argv)
 	unsigned ilast = round(log(opt::minContigLengthEnd) / log(STEP));
 	for (unsigned i = round(log(opt::minContigLength) / log(STEP));
 			i <= ilast; ++i) {
-		unsigned s = pow(STEP, i);
+		unsigned s = pow(STEP, (int)i);
 
 		// Round to 1 figure.
 		double nearestDecade = pow(10, floor(log10(s)));
